Moved utils to folder and renamed to helpers
